Attempt to fix cljdoc
Cljdoc is failing to build the project's docs, because the validation
namespaces require vlad, which is not included in the pom. Before this
patch, vlad was added when needed via a deps.edn alias.

This patch attempts to fix this by making vlad an "optional
dependency"[1] in the pom, instead of an extra dep added via an alias.
In theory, this will do the semantically correct thing: let tooling like
cljdoc see that the project needs vlad, but require that users of f-form
add vlad as their own dep if they intend to use that part of f-form.

The proof will be whether cljdoc is able to analyze the validation
namespaces. Keep an eye on https://cljdoc.org/d/com.github.mainej/f-form
to see.

[1]: https://maven.apache.org/guides/introduction/introduction-to-optional-and-excludes-dependencies.html.